move some type to types file
